Add files via upload
This project is where you write code that can create a special type of barcode called a QR Code. Then, you can also write code that can scan and decode these barcodes and figure out what information is inside them. This is useful for quickly scanning a product in a store and getting information about it.